Division of Beacon Chain into 5 modules
