+++ .env.example
@@ -41,8 +41,62 @@ MODULE_beacon-chain-main_NODES[]=http://login:password@127.0.0.1:1234/
 MODULE_beacon-chain-main_NODES[]=http://login:password@127.0.0.2:1234/
 MODULE_beacon-chain-main_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.3:1234/
 MODULE_beacon-chain-main_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.4:1234/
+MODULE_beacon-chain-main_DEPOSITS[]=http://login:password@127.0.0.5:1234/
+MODULE_beacon-chain-main_DEPOSITS[]=http://login:password@127.0.0.6:1234/
 MODULE_beacon-chain-main_REQUESTER_TIMEOUT=100
 MODULE_beacon-chain-main_REQUESTER_THREADS=12
+
+###################################
+## Attestations Beacon Chain Module
+###################################
+
+MODULES[]=beacon-chain-attestations
+MODULE_beacon-chain-attestations_CLASS=BeaconChainAttestationsModule
+MODULE_beacon-chain-attestations_NODES[]=http://login:password@127.0.0.1:1234/
+MODULE_beacon-chain-attestations_NODES[]=http://login:password@127.0.0.2:1234/
+MODULE_beacon-chain-attestations_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.3:1234/
+MODULE_beacon-chain-attestations_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.4:1234/
+MODULE_beacon-chain-attestations_REQUESTER_TIMEOUT=100
+MODULE_beacon-chain-attestations_REQUESTER_THREADS=12
+
+##################################
+## Withdrawals Beacon Chain Module
+##################################
+
+MODULES[]=beacon-chain-withdrawals
+MODULE_beacon-chain-withdrawals_CLASS=BeaconChainWithdrawalsModule
+MODULE_beacon-chain-withdrawals_NODES[]=http://login:password@127.0.0.1:1234/
+MODULE_beacon-chain-withdrawals_NODES[]=http://login:password@127.0.0.2:1234/
+MODULE_beacon-chain-withdrawals_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.3:1234/
+MODULE_beacon-chain-withdrawals_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.4:1234/
+MODULE_beacon-chain-withdrawals_REQUESTER_TIMEOUT=100
+MODULE_beacon-chain-withdrawals_REQUESTER_THREADS=12
+
+################################
+## Proposals Beacon Chain Module
+################################
+
+MODULES[]=beacon-chain-proposals
+MODULE_beacon-chain-proposals_CLASS=BeaconChainProposalsModule
+MODULE_beacon-chain-proposals_NODES[]=http://login:password@127.0.0.1:1234/
+MODULE_beacon-chain-proposals_NODES[]=http://login:password@127.0.0.2:1234/
+MODULE_beacon-chain-proposals_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.3:1234/
+MODULE_beacon-chain-proposals_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.4:1234/
+MODULE_beacon-chain-proposals_REQUESTER_TIMEOUT=100
+MODULE_beacon-chain-proposals_REQUESTER_THREADS=12
+
+################################
+## Penalties Beacon Chain Module
+################################
+
+MODULES[]=beacon-chain-penalties
+MODULE_beacon-chain-penalties_CLASS=BeaconChainPenaltiesModule
+MODULE_beacon-chain-penalties_NODES[]=http://login:password@127.0.0.1:1234/
+MODULE_beacon-chain-penalties_NODES[]=http://login:password@127.0.0.2:1234/
+MODULE_beacon-chain-penalties_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.3:1234/
+MODULE_beacon-chain-penalties_EXECUTION_LAYER_NODES[]=http://login:password@127.0.0.4:1234/
+MODULE_beacon-chain-penalties_REQUESTER_TIMEOUT=100
+MODULE_beacon-chain-penalties_REQUESTER_THREADS=12
 
 ########################
 ## Trace Ethereum Module
