+++ Modules/BeaconChainAttestationsModule.php
@@ -1,0 +1,29 @@
+<?php declare(strict_types = 1);
+
+/*  Idea (c) 2023 Nikita Zhavoronkov, nikzh@nikzh.com
+ *  Copyright (c) 2023 3xpl developers, 3@3xpl.com, see CONTRIBUTORS.md
+ *  Distributed under the MIT software license, see LICENSE.md  */
+
+/*  This is the Attestations module for Beacon Chain. It requires a Prysm or a Lighthouse node to run.  */
+
+final class BeaconChainAttestationsModule extends BeaconChainLikeAttestationsModule implements Module
+{
+    function initialize()
+    {
+        // CoreModule
+        $this->blockchain = 'beacon-chain';
+        $this->module = 'beacon-chain-attestations';
+        $this->complements = 'beacon-chain-main';
+        $this->is_main = false;
+        $this->first_block_date = '2020-12-01';
+        $this->first_block_id = 0;
+
+        // BeaconChainLikeModule
+        $this->chain_config = [
+            'ALTAIR_FORK_EPOCH' => 74240,
+            'BELLATRIX_FORK_EPOCH' => 144896,
+            'SLOT_PER_EPOCH' => 32,
+            'DELAY' => 2,
+            ];
+    }
+}
