+++ Modules/BeaconChainMainModule.php
@@ -4,7 +4,7 @@
  *  Copyright (c) 2023 3xpl developers, 3@3xpl.com, see CONTRIBUTORS.md
  *  Distributed under the MIT software license, see LICENSE.md  */
 
-/*  This is the main Beacon Chain module. It requires a Prysm or a Lighthouse node to run.  */
+/*  This is the Deposits & Withdrawals (main) module for Beacon Chain. It requires a Prysm or a Lighthouse node to run.  */
 
 final class BeaconChainMainModule extends BeaconChainLikeMainModule implements Module
 {
@@ -16,9 +16,14 @@ final class BeaconChainMainModule extends BeaconChainLikeMainModule implements M
         $this->is_main = true;
         $this->first_block_date = '2020-12-01';
         $this->first_block_id = 0;
-
-        // EVMMainModule
         $this->currency = 'beacon-ethereum'; // We can't use `ethereum` here as this one has a different number of decimals
         $this->currency_details = ['name' => 'Ethereum', 'symbol' => 'ETH', 'decimals' => 9, 'description' => null];
+
+        // BeaconChainLikeModule
+        $this->chain_config = [
+            'BELLATRIX_FORK_EPOCH' => 144896,
+            'SLOT_PER_EPOCH' => 32,
+            'DELAY' => 2,
+        ];
     }
 }
